+++ actix-http/src/client/mod.rs
@@ -10,7 +10,7 @@ mod pool;
 
 pub use self::connection::Connection;
 pub use self::connector::Connector;
-pub use self::error::{ConnectError, InvalidUrl, SendRequestError};
+pub use self::error::{ConnectError, InvalidUrl, SendRequestError, FreezeRequestError};
 pub use self::pool::Protocol;
 
 #[derive(Clone)]